resolved checkmarx issues
