+++ powerflex/helper.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/dell/goscaleio"
+	scaleiotypes "github.com/dell/goscaleio/types/v1"
 )
 
 // getFirstSystem - finds available first system and returns it.
@@ -17,4 +18,27 @@ func getFirstSystem(rc *goscaleio.Client) (*goscaleio.System, error) {
 		return nil, fmt.Errorf("Error in goscaleio FindSystem")
 	}
 	return system, nil
+}
+
+// getNewProtectionDomainEx function to get Protection Domain
+func getNewProtectionDomainEx(c *goscaleio.Client, pdID string, pdName string, href string) (*goscaleio.ProtectionDomain, error) {
+	system, err := getFirstSystem(c)
+	if err != nil {
+		return nil, err
+	}
+	pdr := goscaleio.NewProtectionDomainEx(c, &scaleiotypes.ProtectionDomain{})
+	if pdID != "" {
+		protectionDomain, err := system.FindProtectionDomain(pdID, "", "")
+		pdr.ProtectionDomain = protectionDomain
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		protectionDomain, err := system.FindProtectionDomain("", pdName, "")
+		pdr.ProtectionDomain = protectionDomain
+		if err != nil {
+			return nil, err
+		}
+	}
+	return pdr, nil
 }
